# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Integra Markets API",
    "version": "1.0.0",
    "description": "Commodity news sentiment, narratives and prediction-market divergence.\n\nAuthenticate with `Authorization: Bearer <api_key>`; keys look like `ik_live_...` and are created at https://dashboard.integramarkets.app/account/api\n\n**`sentiment_score` is signed, -1..+1, 0 = neutral** \u2014 that is the field to chart. `confidence` is a separate magnitude and is NOT directional. Before 2026-09-02 the API exposed the confidence value under the name `score`; bearish rows scored higher than bullish ones, so any cached values from before that date must be discarded.\n\nRate limits are per key and are reported on every response as `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Rate` and `X-RateLimit-Reset`. A 429 carries `Retry-After` in seconds \u2014 honour it rather than retrying immediately, because both a monthly allowance and a per-second rate are enforced. Your plan's actual numbers are at https://dashboard.integramarkets.app/account/usage\n\nList endpoints paginate with an opaque `cursor`: when a response carries `has_more: true` it also carries `next_cursor`, which you pass back as `?cursor=`. Do not construct or parse cursors."
  },
  "servers": [
    {
      "url": "https://api.integramarkets.app"
    }
  ],
  "paths": {
    "/v1/commodities": {
      "get": {
        "tags": [
          "sentiment-history"
        ],
        "summary": "List Commodities",
        "description": "Commodities with scored articles in the last 30 days, busiest first.\n\nTHE BUG THIS REPLACES. The previous implementation asked for 10,000 rows of\nentity_mentions with no ORDER BY and no time filter, then took the distinct\nset in Python. Postgres may return any 10,000 rows for such a query, and it\nreturned a block from the archive backfill: this endpoint reported that\n`bitcoin` was the only commodity with data, while the same table held 22 oil\nmentions, 3 gold and 1 copper from the previous 24 hours alone.\n\nA paid endpoint was telling customers the product was empty, and the MCP\ntool built on it was repeating that to their assistant. DISTINCT now happens\nin the database, where it is exact rather than a sample.\n\nCounts ride along because they were free once the grouping existed, and they\nanswer the question behind the question: not \"what can I ask about\" but\n\"what is worth asking about\".",
        "operationId": "list_commodities",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response List Commodities V1 Commodities Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/v1/sentiment/{commodity}/now": {
      "get": {
        "tags": [
          "sentiment-history"
        ],
        "summary": "Sentiment Now",
        "description": "Most recent observed sentiment for `commodity` plus a 24h rolling stat.",
        "operationId": "get_sentiment_now",
        "parameters": [
          {
            "name": "commodity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Commodity"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Sentiment Now V1 Sentiment  Commodity  Now Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/v1/sentiment/{commodity}/history": {
      "get": {
        "tags": [
          "sentiment-history"
        ],
        "summary": "Sentiment History",
        "description": "Time-series of individual scored documents for `commodity`.",
        "operationId": "get_sentiment_history",
        "parameters": [
          {
            "name": "commodity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Commodity"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO 8601 UTC timestamp",
              "title": "From"
            },
            "description": "ISO 8601 UTC timestamp"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO 8601 UTC timestamp",
              "title": "To"
            },
            "description": "ISO 8601 UTC timestamp"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Sentiment History V1 Sentiment  Commodity  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/v1/sentiment/{commodity}/daily": {
      "get": {
        "tags": [
          "sentiment-history"
        ],
        "summary": "Sentiment Daily",
        "description": "Daily aggregates for the last N days, computed on-the-fly.",
        "operationId": "get_sentiment_daily",
        "parameters": [
          {
            "name": "commodity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Commodity"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "default": 30,
              "title": "Days"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Sentiment Daily V1 Sentiment  Commodity  Daily Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/v1/markets/overlay": {
      "get": {
        "tags": [
          "sentiment-history"
        ],
        "summary": "Markets Overlay",
        "description": "Resolved prediction markets with linked news-sentiment context.\n\nFor the beta this returns the resolved markets enriched with the\naverage news sentiment over the market's lifetime. Full per-snapshot\noverlay arrives once the nightly aggregation job is wired up; this\nendpoint already returns the shape that job will populate.",
        "operationId": "get_market_overlay",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "kalshi",
              "title": "Provider"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "market status filter applied to raw_payload",
              "default": "settled",
              "title": "Status"
            },
            "description": "market status filter applied to raw_payload"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Markets Overlay V1 Markets Overlay Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/v1/sentiment": {
      "get": {
        "tags": [
          "public-v1"
        ],
        "summary": "Sentiment",
        "operationId": "get_sentiment",
        "parameters": [
          {
            "name": "commodity",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Canonical commodity name, e.g. 'oil', 'gas', 'copper'. This is matched against entity_mentions.entity, which stores the name the sentiment engine normalises to \u2014 not a market ticker, so 'brent' matches nothing.",
              "title": "Commodity"
            },
            "description": "Canonical commodity name, e.g. 'oil', 'gas', 'copper'. This is matched against entity_mentions.entity, which stores the name the sentiment engine normalises to \u2014 not a market ticker, so 'brent' matches nothing."
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "One of 24h, 7d, 30d, 90d",
              "default": "7d",
              "title": "Window"
            },
            "description": "One of 24h, 7d, 30d, 90d"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Sentiment V1 Sentiment Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "An Integra API key, e.g. ik_live_..., sent as `Authorization: Bearer <key>`. Keys carry the scopes of the subscription that created them, resolved live per request; a valid key with no active entitlement receives 403 rather than 401. Manage keys at https://dashboard.integramarkets.app/account/api"
      }
    }
  }
}